Updated the app title; turned off autocomplete for form inputs and Fixed a bug with headless ui tabs responding to changes in mst data
